+++ src/constants.rs
@@ -1,0 +1,2 @@
+/// The base URI to the API.
+pub const API_URL: &'static str = "https://api.darksky.net";